+++ src/main/java/shit/zen/modules/impl/movement/TargetStrafe.java
@@ -19,8 +19,8 @@ extends Module {
     public static TargetStrafe INSTANCE;
     private final Timer collisionTimer = new Timer();
     private final BooleanSetting smartStrafe = new BooleanSetting("Jump Key Only", true);
-    private final NumberSetting range = new NumberSetting("Range", Float.valueOf(0.5f), Float.valueOf(0.1f), Float.valueOf(2.0f), Float.valueOf(0.1f));
-    private final NumberSetting switchDelay = new NumberSetting("Switch Delay", Integer.valueOf(1000), Integer.valueOf(100), Integer.valueOf(5000), Integer.valueOf(100));
+    private final NumberSetting range = new NumberSetting("Range", 0.5f, 0.1f, 2.0f, 0.1f);
+    private final NumberSetting switchDelay = new NumberSetting("Switch Delay", 1000, 100, 5000, 100);
     public static int strafeDirectionSign;
     public static Entity strafeTarget;
     private final Timer switchTimer = new Timer();
