+++ src/main/java/shit/zen/modules/impl/player/AutoMLG.java
@@ -23,8 +23,8 @@ import shit.zen.event.EventTarget;
 public class AutoMLG
 extends Module {
     public static AutoMLG INSTANCE;
-    private final NumberSetting triggerDistanceSetting = new NumberSetting("Fall distance", Float.valueOf(3.0f), Float.valueOf(1.0f), Float.valueOf(10.0f), Float.valueOf(0.1f));
-    private final NumberSetting predictTicksSetting = new NumberSetting("Predict Ticks", Float.valueOf(2.0f), Float.valueOf(1.0f), Float.valueOf(5.0f), Float.valueOf(1.0f));
+    private final NumberSetting triggerDistanceSetting = new NumberSetting("Fall distance", 3.0f, 1.0f, 10.0f, 0.1f);
+    private final NumberSetting predictTicksSetting = new NumberSetting("Predict Ticks", 2.0f, 1.0f, 5.0f, 1.0f);
     private final BooleanSetting solidCheckSetting = new BooleanSetting("Solid check", true);
     private final BooleanSetting recoverySetting = new BooleanSetting("Recorvey", true);
     public Rotation targetRotation = null;
